ci: use OIDC for publishing releases (#27)
This switches us to using [trusted
publishing](https://docs.npmjs.com/trusted-publishers) aka OIDC for
authenticating with the NPM registry to publish releases, rather than
long-lived access tokens.

This is more secure and will become [all but
required](https://github.blog/changelog/2025-09-29-strengthening-npm-security-important-changes-to-authentication-and-token-management/)
for CI-based releases, as publishing tokens will have a max expiration
of 90 days.

Because of the magic that is OIDC this should "just work" 🤞